# backend/openapi_audit.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Victor Institutional Logic",
                 "version":  "0.1.0"
             },
    "paths":  {
                  "/api/v1/signals/":  {
                                           "get":  {
                                                       "tags":  [
                                                                    "signals"
                                                                ],
                                                       "summary":  "Read Signals",
                                                       "operationId":  "read_signals_api_v1_signals__get",
                                                       "parameters":  [
                                                                          {
                                                                              "name":  "skip",
                                                                              "in":  "query",
                                                                              "required":  false,
                                                                              "schema":  {
                                                                                             "type":  "integer",
                                                                                             "default":  0,
                                                                                             "title":  "Skip"
                                                                                         }
                                                                          },
                                                                          {
                                                                              "name":  "limit",
                                                                              "in":  "query",
                                                                              "required":  false,
                                                                              "schema":  {
                                                                                             "type":  "integer",
                                                                                             "default":  100,
                                                                                             "title":  "Limit"
                                                                                         }
                                                                          }
                                                                      ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "Successful Response",
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "type":  "array",
                                                                                                                                             "items":  {
                                                                                                                                                           "$ref":  "#/components/schemas/Signal"
                                                                                                                                                       },
                                                                                                                                             "title":  "Response Read Signals Api V1 Signals  Get"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 },
                                                                         "422":  {
                                                                                     "description":  "Validation Error",
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 }
                                                                     }
                                                   },
                                           "post":  {
                                                        "tags":  [
                                                                     "signals"
                                                                 ],
                                                        "summary":  "Create Signal",
                                                        "operationId":  "create_signal_api_v1_signals__post",
                                                        "requestBody":  {
                                                                            "required":  true,
                                                                            "content":  {
                                                                                            "application/json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/SignalCreate"
                                                                                                                                }
                                                                                                                 }
                                                                                        }
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Successful Response",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/Signal"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  },
                                                                          "422":  {
                                                                                      "description":  "Validation Error",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  }
                                                                      }
                                                    }
                                       },
...
              },
    "components":  {
                       "schemas":  {
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "Signal":  {
                                                      "properties":  {
                                                                         "symbol":  {
                                                                                        "type":  "string",
                                                                                        "title":  "Symbol"
                                                                                    },
                                                                         "signal_type":  {
                                                                                             "$ref":  "#/components/schemas/SignalType"
                                                                                         },
                                                                         "price":  {
                                                                                       "type":  "number",
                                                                                       "title":  "Price"
                                                                                   },
                                                                         "status":  {
                                                                                        "anyOf":  [
                                                                                                      {
                                                                                                          "$ref":  "#/components/schemas/SignalStatus"
                                                                                                      },
                                                                                                      {
                                                                                                          "type":  "null"
                                                                                                      }
                                                                                                  ],
                                                                                        "default":  "PENDING"
                                                                                    },
                                                                         "id":  {
                                                                                    "type":  "integer",
                                                                                    "title":  "Id"
                                                                                },
                                                                         "created_at":  {
                                                                                            "type":  "string",
                                                                                            "format":  "date-time",
                                                                                            "title":  "Created At"
                                                                                        }
                                                                     },
                                                      "type":  "object",
                                                      "required":  [
                                                                       "symbol",
                                                                       "signal_type",
                                                                       "price",
                                                                       "id",
                                                                       "created_at"
                                                                   ],
                                                      "title":  "Signal"
                                                  },
                                       "SignalCreate":  {
                                                            "properties":  {
                                                                               "symbol":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Symbol"
                                                                                          },
                                                                               "signal_type":  {
                                                                                                   "$ref":  "#/components/schemas/SignalType"
                                                                                               },
                                                                               "price":  {
                                                                                             "type":  "number",
                                                                                             "title":  "Price"
                                                                                         },
                                                                               "status":  {
                                                                                              "anyOf":  [
                                                                                                            {
                                                                                                                "$ref":  "#/components/schemas/SignalStatus"
                                                                                                            },
                                                                                                            {
                                                                                                                "type":  "null"
                                                                                                            }
                                                                                                        ],
                                                                                              "default":  "PENDING"
                                                                                          }
                                                                           },
                                                            "type":  "object",
                                                            "required":  [
                                                                             "symbol",
                                                                             "signal_type",
                                                                             "price"
                                                                         ],
                                                            "title":  "SignalCreate"
                                                        },
...
                                       "SignalType":  {
                                                          "type":  "string",
                                                          "enum":  [
                                                                       "BUY",
                                                                       "SELL"
                                                                   ],
                                                          "title":  "SignalType"
                                                      },
...
                                   }
                   }
}